Convert "epoch_time" into DateTime
Previously "epoch_time" was passed in as a string. This creates problem
for gpu usage, because strings are not isbitstypes. This commit adds "datetime"
as a valtype.
It is also changes "epoch_time" to a datetime in the toml file, which natively supports
DateTimes. Users can still pass the "epoch_time" as a string. which is converted
to a DateTime.